+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 0.2.3 / ####-##-##
 ==================
 
+  * Click now supports index and clicks all elements
+
 0.2.2 / 2018-02-01
 ==================